isAuthenticated 2 y restricción permisos en borrar reseña

+++ src/main/java/es/uma/taw/arkhammovies/controller/CommonController.java
@@ -4,9 +4,11 @@ import es.uma.taw.arkhammovies.dto.*;
 import es.uma.taw.arkhammovies.service.*;
 import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.util.Collections;
@@ -191,6 +193,8 @@ public class CommonController extends BaseController{
 
     @GetMapping("/flipLike")
     public String flipLike(HttpSession session, Model model, @ModelAttribute("movieId") Integer movieId, @RequestParam Boolean tipo) {
+        if (!isAuthenticated(session)) return "redirect:/user/login";
+
         UserDTO user = (UserDTO) session.getAttribute("user");
         MovieDTO movie = movieService.findMovie(movieId);
 
@@ -207,6 +211,8 @@ public class CommonController extends BaseController{
 
     @GetMapping("/flipSave")
     public String flipSave(HttpSession session, Model model, @ModelAttribute("movieId") Integer movieId,@RequestParam Boolean tipo) {
+        if (!isAuthenticated(session)) return "redirect:/user/login";
+
         UserDTO user = (UserDTO) session.getAttribute("user");
         MovieDTO movie = movieService.findMovie(movieId);
 
@@ -227,11 +233,11 @@ public class CommonController extends BaseController{
                               @RequestParam("movieId") Integer movieId,
                               @RequestParam(required = false, name = "score") Integer score,
                               @RequestParam("reviewText") String review) {
+        if (!isAuthenticated(session)) return "redirect:/user/login";
+
         UserDTO user = (UserDTO) session.getAttribute("user");
-        if(user == null) {
-            return "redirect:/user/login";
-        }
-        else if (score != null && !review.isEmpty()){
+
+        if (score != null && !review.isEmpty()){
             MovieDTO movie = movieService.findMovie(movieId);
             reviewService.addReview(movie.getId(), user.getId(), score, review);
         } else { // Atributos para redirección
@@ -242,7 +248,16 @@ public class CommonController extends BaseController{
     }
 
     @GetMapping("/removeReview")
-    public String doRemoveReview(@RequestParam("movieId") Integer movieId, @RequestParam("userId") Integer userId) {
+    public String doRemoveReview(@RequestParam("movieId") Integer movieId, @RequestParam("userId") Integer userId,
+                                 HttpSession session) {
+        if (!isAuthenticated(session)) return "redirect:/user/login";
+
+        UserDTO user = (UserDTO) session.getAttribute("user");
+
+        if (user.getRole() >= 2 && !user.getId().equals(userId)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
         this.reviewService.removeById(movieId, userId);
         return "redirect:/movies/movie?id=" + movieId;
     }
